services/manager/app/requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
opensearch-py==2.6.0
pyyaml==6.0.2
netaddr==1.3.0

authlib==1.3.1
python-jose==3.3.0
